requirements.txt
fastapi
uvicorn
requests
python-dotenv
pandas
numpy
google-generativeai
